requirements.txt
omxplayer-wrapper==0.2.5
RPi.GPIO==0.6.3